+++ settingrepo-node/src/main/java/com/dwarfeng/settingrepo/node/handler/LauncherSettingHandler.java
@@ -10,7 +10,14 @@ public class LauncherSettingHandler implements Handler {
     @Value("${launcher.reset_formatter_support}")
     private boolean resetFormatterSupport;
 
+    @Value("${launcher.start_reset_delay}")
+    private long startResetDelay;
+
     public boolean isResetFormatterSupport() {
         return resetFormatterSupport;
+    }
+
+    public long getStartResetDelay() {
+        return startResetDelay;
     }
 }